Update Ikea_E1745.md
I added the "press 5 times to get in pairing mode" because holding the button for 10 seconds did not work for me. Might be firmware-version dependent?
According to this source (https://community.smartthings.com/t/ikea-tradfri-motion-sensor-paring-issues/188209), the 10-sec-press is used to directly pair the sensor to a light bulb, so maybe it should be removed here?!

"Holding the button for 10 seconds is the pairing process to a bulb. 5 quick presses factory resets it then it enters pairing mode for the hub."